Cleanup test/e2e/upgrades directory
