+++ test/e2e/upgrades/node/configmaps.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package upgrades
+package node
 
 import (
 	"context"
@@ -24,84 +24,89 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/kubernetes/test/e2e/upgrades"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 
 	"github.com/onsi/ginkgo"
 )
 
-// SecretUpgradeTest test that a secret is available before and after
+// ConfigMapUpgradeTest tests that a ConfigMap is available before and after
 // a cluster upgrade.
-type SecretUpgradeTest struct {
-	secret *v1.Secret
+type ConfigMapUpgradeTest struct {
+	configMap *v1.ConfigMap
 }
 
 // Name returns the tracking name of the test.
-func (SecretUpgradeTest) Name() string { return "[sig-storage] [sig-api-machinery] secret-upgrade" }
+func (ConfigMapUpgradeTest) Name() string {
+	return "[sig-storage] [sig-api-machinery] configmap-upgrade"
+}
 
-// Setup creates a secret and then verifies that a pod can consume it.
-func (t *SecretUpgradeTest) Setup(f *framework.Framework) {
-	secretName := "upgrade-secret"
+// Setup creates a ConfigMap and then verifies that a pod can consume it.
+func (t *ConfigMapUpgradeTest) Setup(f *framework.Framework) {
+	configMapName := "upgrade-configmap"
 
 	ns := f.Namespace
 
-	t.secret = &v1.Secret{
+	t.configMap = &v1.ConfigMap{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: ns.Name,
-			Name:      secretName,
+			Name:      configMapName,
 		},
-		Data: map[string][]byte{
-			"data": []byte("keep it secret"),
+		Data: map[string]string{
+			"data": "some configmap data",
 		},
 	}
 
-	ginkgo.By("Creating a secret")
+	ginkgo.By("Creating a ConfigMap")
 	var err error
-	if t.secret, err = f.ClientSet.CoreV1().Secrets(ns.Name).Create(context.TODO(), t.secret, metav1.CreateOptions{}); err != nil {
-		framework.Failf("unable to create test secret %s: %v", t.secret.Name, err)
+	if t.configMap, err = f.ClientSet.CoreV1().ConfigMaps(ns.Name).Create(context.TODO(), t.configMap, metav1.CreateOptions{}); err != nil {
+		framework.Failf("unable to create test ConfigMap %s: %v", t.configMap.Name, err)
 	}
 
-	ginkgo.By("Making sure the secret is consumable")
+	ginkgo.By("Making sure the ConfigMap is consumable")
 	t.testPod(f)
 }
 
 // Test waits for the upgrade to complete, and then verifies that a
-// pod can still consume the secret.
-func (t *SecretUpgradeTest) Test(f *framework.Framework, done <-chan struct{}, upgrade UpgradeType) {
+// pod can still consume the ConfigMap.
+func (t *ConfigMapUpgradeTest) Test(f *framework.Framework, done <-chan struct{}, upgrade upgrades.UpgradeType) {
 	<-done
-	ginkgo.By("Consuming the secret after upgrade")
+	ginkgo.By("Consuming the ConfigMap after upgrade")
 	t.testPod(f)
 }
 
 // Teardown cleans up any remaining resources.
-func (t *SecretUpgradeTest) Teardown(f *framework.Framework) {
+func (t *ConfigMapUpgradeTest) Teardown(f *framework.Framework) {
 	// rely on the namespace deletion to clean up everything
 }
 
-// testPod creates a pod that consumes a secret and prints it out. The
+// testPod creates a pod that consumes a ConfigMap and prints it out. The
 // output is then verified.
-func (t *SecretUpgradeTest) testPod(f *framework.Framework) {
-	volumeName := "secret-volume"
-	volumeMountPath := "/etc/secret-volume"
+func (t *ConfigMapUpgradeTest) testPod(f *framework.Framework) {
+	volumeName := "configmap-volume"
+	volumeMountPath := "/etc/configmap-volume"
 
 	pod := &v1.Pod{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "pod-secrets-" + string(uuid.NewUUID()),
-			Namespace: t.secret.ObjectMeta.Namespace,
+			Name:      "pod-configmap-" + string(uuid.NewUUID()),
+			Namespace: t.configMap.ObjectMeta.Namespace,
 		},
 		Spec: v1.PodSpec{
 			Volumes: []v1.Volume{
 				{
 					Name: volumeName,
 					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: t.secret.ObjectMeta.Name,
+						ConfigMap: &v1.ConfigMapVolumeSource{
+							LocalObjectReference: v1.LocalObjectReference{
+								Name: t.configMap.ObjectMeta.Name,
+							},
 						},
 					},
 				},
 			},
 			Containers: []v1.Container{
 				{
-					Name:  "secret-volume-test",
+					Name:  "configmap-volume-test",
 					Image: imageutils.GetE2EImage(imageutils.Agnhost),
 					Args: []string{
 						"mounttest",
@@ -116,16 +121,16 @@ func (t *SecretUpgradeTest) testPod(f *framework.Framework) {
 					},
 				},
 				{
-					Name:    "secret-env-test",
+					Name:    "configmap-env-test",
 					Image:   imageutils.GetE2EImage(imageutils.BusyBox),
 					Command: []string{"sh", "-c", "env"},
 					Env: []v1.EnvVar{
 						{
-							Name: "SECRET_DATA",
+							Name: "CONFIGMAP_DATA",
 							ValueFrom: &v1.EnvVarSource{
-								SecretKeyRef: &v1.SecretKeySelector{
+								ConfigMapKeyRef: &v1.ConfigMapKeySelector{
 									LocalObjectReference: v1.LocalObjectReference{
-										Name: t.secret.ObjectMeta.Name,
+										Name: t.configMap.ObjectMeta.Name,
 									},
 									Key: "data",
 								},
@@ -139,12 +144,11 @@ func (t *SecretUpgradeTest) testPod(f *framework.Framework) {
 	}
 
 	expectedOutput := []string{
-		"content of file \"/etc/secret-volume/data\": keep it secret",
-		"mode of file \"/etc/secret-volume/data\": -rw-r--r--",
+		"content of file \"/etc/configmap-volume/data\": some configmap data",
+		"mode of file \"/etc/configmap-volume/data\": -rw-r--r--",
 	}
+	f.TestContainerOutput("volume consume configmap", pod, 0, expectedOutput)
 
-	f.TestContainerOutput("volume consume secrets", pod, 0, expectedOutput)
-
-	expectedOutput = []string{"SECRET_DATA=keep it secret"}
-	f.TestContainerOutput("env consume secrets", pod, 1, expectedOutput)
+	expectedOutput = []string{"CONFIGMAP_DATA=some configmap data"}
+	f.TestContainerOutput("env consume configmap", pod, 1, expectedOutput)
 }
